importer: purge all PGDUMP and MYSQLDUMP related code
All this code is no longer used. Note that there are still some more
things to clean up (like we can assume that no tables are new, and we
don't need to support creating new schemas in IMPORT either), but all
those things will be addressed separately. This commit focuses only on
removing the related workload readers as well as the roachpb formats.

Release note: None
