+++ pkg/cmd/cr2pg/sqlstream/stream.go
@@ -17,8 +17,6 @@ import (
 	"github.com/cockroachdb/errors"
 )
 
-// Modified from importer/read_import_pgdump.go.
-
 // Stream streams an io.Reader into tree.Statements.
 type Stream struct {
 	scan *bufio.Scanner
