+++ pkg/roachpb/io-formats.proto
@@ -15,19 +15,17 @@ message IOFileFormat {
     Unknown = 0;
     CSV = 1;
     MysqlOutfile = 2;
-    Mysqldump = 3;
     PgCopy = 4;
-    PgDump = 5;
     Avro = 6;
     Parquet = 7;
+
+    reserved 3, 5;
   }
 
   optional FileFormat format = 1 [(gogoproto.nullable) = false];
   optional CSVOptions csv = 2 [(gogoproto.nullable) = false];
   optional MySQLOutfileOptions mysql_out = 3 [(gogoproto.nullable) = false];
   optional PgCopyOptions pg_copy = 4 [(gogoproto.nullable) = false];
-  optional MysqldumpOptions mysql_dump = 9 [(gogoproto.nullable) = false];
-  optional PgDumpOptions pg_dump = 6 [(gogoproto.nullable) = false];
   optional AvroOptions avro = 8 [(gogoproto.nullable) = false];
   optional ParquetOptions parquet = 10 [(gogoproto.nullable) = false];
 
@@ -41,6 +39,8 @@ message IOFileFormat {
   optional Compression compression = 5 [(gogoproto.nullable) = false];
   // If true, don't abort on failures but instead save the offending row and keep on.
   optional bool save_rejected = 7 [(gogoproto.nullable) = false];
+
+  reserved 6, 9;
 }
 
 
@@ -101,29 +101,6 @@ message PgCopyOptions {
   optional string null = 2 [(gogoproto.nullable) = false];
   // maxRowSize is the maximum row size
   optional int32 maxRowSize = 3 [(gogoproto.nullable) = false];
-}
-
-// PgDumpOptions describe the format of postgresql's pg_dump.
-message PgDumpOptions {
-  // maxRowSize is the maximum row size
-  optional int32 maxRowSize = 1 [(gogoproto.nullable) = false];
-  // Indicates the number of rows to import per table.
-  // Must be a non-zero positive number. 
-  optional int64 row_limit = 2 [(gogoproto.nullable) = false];
-  // Indicates if all unparseable and parseable, but unimplemented PGDUMP stmts
-  // should be ignored during IMPORT.
-  optional bool ignore_unsupported = 3 [(gogoproto.nullable) = false];
-  // Points to the destination where unsupported statements during a PGDUMP
-  // import should be logged. This can only be used when ignore_unsupported is
-  // specified, otherwise the IMPORT errors out on encountering an unsupported
-  // stmt.
-  optional string ignore_unsupported_log = 4 [(gogoproto.nullable) = false];
-}
-
-message MysqldumpOptions {
-  // Indicates the number of rows to import per table.
-  // Must be a non-zero positive number. 
-  optional int64 row_limit = 1 [(gogoproto.nullable) = false];
 }
 
 message AvroOptions {
