+++ pkg/sql/importer/BUILD.bazel
@@ -17,10 +17,8 @@ go_library(
         "read_import_avro.go",
         "read_import_base.go",
         "read_import_csv.go",
-        "read_import_mysql.go",
         "read_import_mysqlout.go",
         "read_import_pgcopy.go",
-        "read_import_pgdump.go",
         "read_import_workload.go",
         "rollback_job.go",
     ],
@@ -31,7 +29,6 @@ go_library(
         "//pkg/clusterversion",
         "//pkg/col/coldata",
         "//pkg/crosscluster",
-        "//pkg/docs",
         "//pkg/featureflag",
         "//pkg/jobs",
         "//pkg/jobs/ingeststopped",
@@ -51,7 +48,6 @@ go_library(
         "//pkg/sql/catalog/catpb",
         "//pkg/sql/catalog/colinfo",
         "//pkg/sql/catalog/dbdesc",
-        "//pkg/sql/catalog/descidgen",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/ingesting",
@@ -70,10 +66,8 @@ go_library(
         "//pkg/sql/isql",
         "//pkg/sql/lexbase",
         "//pkg/sql/opt/memo",
-        "//pkg/sql/parser",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
-        "//pkg/sql/pgwire/pgnotice",
         "//pkg/sql/physicalplan",
         "//pkg/sql/privilege",
         "//pkg/sql/row",
@@ -96,7 +90,6 @@ go_library(
         "//pkg/util/errorutil/unimplemented",
         "//pkg/util/hlc",
         "//pkg/util/humanizeutil",
-        "//pkg/util/intsets",
         "//pkg/util/ioctx",
         "//pkg/util/log",
         "//pkg/util/log/eventpb",
@@ -115,8 +108,6 @@ go_library(
         "@com_github_cockroachdb_redact//:redact",
         "@com_github_lib_pq//oid",
         "@com_github_linkedin_goavro_v2//:goavro",
-        "@io_vitess_vitess//go/sqltypes",
-        "@io_vitess_vitess//go/vt/sqlparser",
     ],
 )
