+++ pkg/sql/importer/import_planning.go
@@ -6,7 +6,6 @@
 package importer
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"math"
@@ -17,12 +16,10 @@ import (
 	"strings"
 
 	"github.com/cockroachdb/cockroach/pkg/cloud"
-	"github.com/cockroachdb/cockroach/pkg/docs"
 	"github.com/cockroachdb/cockroach/pkg/featureflag"
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
-	"github.com/cockroachdb/cockroach/pkg/security/username"
 	"github.com/cockroachdb/cockroach/pkg/server/telemetry"
 	"github.com/cockroachdb/cockroach/pkg/settings"
 	"github.com/cockroachdb/cockroach/pkg/sql"
@@ -38,7 +35,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/isql"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgnotice"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/idxtype"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
@@ -89,15 +85,6 @@ const (
 	// as either an inline JSON schema, or an external schema URI.
 	avroSchema    = "schema"
 	avroSchemaURI = "schema_uri"
-
-	pgDumpIgnoreAllUnsupported     = "ignore_unsupported_statements"
-	pgDumpIgnoreShuntFileDest      = "log_ignored_statements"
-	pgDumpUnsupportedSchemaStmtLog = "unsupported_schema_stmts"
-	pgDumpUnsupportedDataStmtLog   = "unsupported_data_stmts"
-
-	// statusImportBundleParseSchema indicates to the user that a bundle format
-	// schema is being parsed
-	statusImportBundleParseSchema jobs.StatusMessage = "parsing schema on Import Bundle"
 )
 
 var importOptionExpectValues = map[string]exprutil.KVStringOptValidate{
@@ -129,12 +116,7 @@ var importOptionExpectValues = map[string]exprutil.KVStringOptValidate{
 	avroRecordsSeparatedBy: exprutil.KVStringOptRequireValue,
 	avroBinRecords:         exprutil.KVStringOptRequireNoValue,
 	avroJSONRecords:        exprutil.KVStringOptRequireNoValue,
-
-	pgDumpIgnoreAllUnsupported: exprutil.KVStringOptRequireNoValue,
-	pgDumpIgnoreShuntFileDest:  exprutil.KVStringOptRequireValue,
 }
-
-var pgDumpMaxLoggedStmts = 1024
 
 func makeStringSet(opts ...string) map[string]struct{} {
 	res := make(map[string]struct{}, len(opts))
@@ -164,12 +146,7 @@ var mysqlOutAllowedOptions = makeStringSet(
 	mysqlOutfileEscape, csvNullIf, csvSkip, csvRowLimit,
 )
 
-var (
-	mysqlDumpAllowedOptions = makeStringSet(importOptionSkipFKs, csvRowLimit)
-	pgCopyAllowedOptions    = makeStringSet(pgCopyDelimiter, pgCopyNull, optMaxRowSize)
-	pgDumpAllowedOptions    = makeStringSet(optMaxRowSize, importOptionSkipFKs, csvRowLimit,
-		pgDumpIgnoreAllUnsupported, pgDumpIgnoreShuntFileDest)
-)
+var pgCopyAllowedOptions = makeStringSet(pgCopyDelimiter, pgCopyNull, optMaxRowSize)
 
 // DROP is required because the target table needs to be take offline during
 // IMPORT INTO.
@@ -341,15 +318,6 @@ func importPlanHook(
 	importStmt, ok := stmt.(*tree.Import)
 	if !ok {
 		return nil, nil, false, nil
-	}
-
-	switch f := strings.ToUpper(importStmt.FileFormat); f {
-	case "PGDUMP", "MYSQLDUMP":
-		p.BufferClientNotice(ctx, pgnotice.Newf(
-			"IMPORT %s has been deprecated in 23.1, and will be removed in a future version. See %s for alternatives.",
-			redact.SafeString(f),
-			redact.SafeString(docs.URL("migration-overview")),
-		))
 	}
 
 	addToFileFormatTelemetry(importStmt.FileFormat, "attempted")
@@ -603,21 +571,6 @@ func importPlanHook(
 				}
 				format.MysqlOut.RowLimit = int64(rowLimit)
 			}
-		case "MYSQLDUMP":
-			if err = validateFormatOptions(importStmt.FileFormat, opts, mysqlDumpAllowedOptions); err != nil {
-				return err
-			}
-			format.Format = roachpb.IOFileFormat_Mysqldump
-			if override, ok := opts[csvRowLimit]; ok {
-				rowLimit, err := strconv.Atoi(override)
-				if err != nil {
-					return pgerror.Wrapf(err, pgcode.Syntax, "invalid numeric %s value", csvRowLimit)
-				}
-				if rowLimit <= 0 {
-					return pgerror.Newf(pgcode.Syntax, "%s must be > 0", csvRowLimit)
-				}
-				format.MysqlDump.RowLimit = int64(rowLimit)
-			}
 		case "PGCOPY":
 			if err = validateFormatOptions(importStmt.FileFormat, opts, pgCopyAllowedOptions); err != nil {
 				return err
@@ -649,44 +602,6 @@ func importPlanHook(
 				maxRowSize = int32(sz)
 			}
 			format.PgCopy.MaxRowSize = maxRowSize
-		case "PGDUMP":
-			if err = validateFormatOptions(importStmt.FileFormat, opts, pgDumpAllowedOptions); err != nil {
-				return err
-			}
-			format.Format = roachpb.IOFileFormat_PgDump
-			maxRowSize := int32(defaultScanBuffer)
-			if override, ok := opts[optMaxRowSize]; ok {
-				sz, err := humanizeutil.ParseBytes(override)
-				if err != nil {
-					return err
-				}
-				if sz < 1 || sz > math.MaxInt32 {
-					return errors.Errorf("%d out of range: %d", maxRowSize, sz)
-				}
-				maxRowSize = int32(sz)
-			}
-			format.PgDump.MaxRowSize = maxRowSize
-			if _, ok := opts[pgDumpIgnoreAllUnsupported]; ok {
-				format.PgDump.IgnoreUnsupported = true
-			}
-
-			if dest, ok := opts[pgDumpIgnoreShuntFileDest]; ok {
-				if !format.PgDump.IgnoreUnsupported {
-					return errors.New("cannot log unsupported PGDUMP stmts without `ignore_unsupported_statements` option")
-				}
-				format.PgDump.IgnoreUnsupportedLog = dest
-			}
-
-			if override, ok := opts[csvRowLimit]; ok {
-				rowLimit, err := strconv.Atoi(override)
-				if err != nil {
-					return pgerror.Wrapf(err, pgcode.Syntax, "invalid numeric %s value", csvRowLimit)
-				}
-				if rowLimit <= 0 {
-					return pgerror.Newf(pgcode.Syntax, "%s must be > 0", csvRowLimit)
-				}
-				format.PgDump.RowLimit = int64(rowLimit)
-			}
 		case "AVRO":
 			if err = validateFormatOptions(importStmt.FileFormat, opts, avroAllowedOptions); err != nil {
 				return err
@@ -1031,115 +946,6 @@ func parseAvroOptions(
 			format.Avro.MaxRecordSize = int32(sz)
 		}
 	}
-	return nil
-}
-
-type loggerKind int
-
-const (
-	schemaParsing loggerKind = iota
-	dataIngestion
-)
-
-// unsupportedStmtLogger is responsible for handling unsupported PGDUMP SQL
-// statements seen during the import.
-type unsupportedStmtLogger struct {
-	ctx   context.Context
-	user  username.SQLUsername
-	jobID int64
-
-	// Values are initialized based on the options specified in the IMPORT PGDUMP
-	// stmt.
-	ignoreUnsupported        bool
-	ignoreUnsupportedLogDest string
-	externalStorage          cloud.ExternalStorageFactory
-
-	// logBuffer holds the string to be flushed to the ignoreUnsupportedLogDest.
-	logBuffer       *bytes.Buffer
-	numIgnoredStmts int
-
-	// Incremented every time the logger flushes. It is used as the suffix of the
-	// log file written to external storage.
-	flushCount int
-
-	loggerType loggerKind
-}
-
-func makeUnsupportedStmtLogger(
-	ctx context.Context,
-	user username.SQLUsername,
-	jobID int64,
-	ignoreUnsupported bool,
-	unsupportedLogDest string,
-	loggerType loggerKind,
-	externalStorage cloud.ExternalStorageFactory,
-) *unsupportedStmtLogger {
-	return &unsupportedStmtLogger{
-		ctx:                      ctx,
-		user:                     user,
-		jobID:                    jobID,
-		ignoreUnsupported:        ignoreUnsupported,
-		ignoreUnsupportedLogDest: unsupportedLogDest,
-		loggerType:               loggerType,
-		logBuffer:                new(bytes.Buffer),
-		externalStorage:          externalStorage,
-	}
-}
-
-func (u *unsupportedStmtLogger) log(logLine string, isParseError bool) error {
-	// We have already logged parse errors during the schema ingestion phase, so
-	// skip them to avoid duplicate entries.
-	skipLoggingParseErr := isParseError && u.loggerType == dataIngestion
-	if u.ignoreUnsupportedLogDest == "" || skipLoggingParseErr {
-		return nil
-	}
-
-	// Flush to a file if we have hit the max size of our buffer.
-	if u.numIgnoredStmts >= pgDumpMaxLoggedStmts {
-		err := u.flush()
-		if err != nil {
-			return err
-		}
-	}
-
-	if isParseError {
-		logLine = fmt.Sprintf("%s: could not be parsed\n", logLine)
-	} else {
-		logLine = fmt.Sprintf("%s: unsupported by IMPORT\n", logLine)
-	}
-	u.logBuffer.Write([]byte(logLine))
-	u.numIgnoredStmts++
-	return nil
-}
-
-func (u *unsupportedStmtLogger) flush() error {
-	if u.ignoreUnsupportedLogDest == "" {
-		return nil
-	}
-
-	conf, err := cloud.ExternalStorageConfFromURI(u.ignoreUnsupportedLogDest, u.user)
-	if err != nil {
-		return errors.Wrap(err, "failed to log unsupported stmts during IMPORT PGDUMP")
-	}
-	var s cloud.ExternalStorage
-	if s, err = u.externalStorage(u.ctx, conf); err != nil {
-		return errors.New("failed to log unsupported stmts during IMPORT PGDUMP")
-	}
-	defer s.Close()
-
-	logFileName := fmt.Sprintf("import%d", u.jobID)
-	if u.loggerType == dataIngestion {
-		logFileName = path.Join(logFileName, pgDumpUnsupportedDataStmtLog, fmt.Sprintf("%d.log", u.flushCount))
-	} else {
-		logFileName = path.Join(logFileName, pgDumpUnsupportedSchemaStmtLog, fmt.Sprintf("%d.log", u.flushCount))
-	}
-	err = cloud.WriteFile(u.ctx, s, logFileName, bytes.NewReader(u.logBuffer.Bytes()))
-	if err != nil {
-		return errors.Wrap(err, "failed to log unsupported stmts to log during IMPORT PGDUMP")
-	}
-	u.flushCount++
-	u.numIgnoredStmts = 0
-	u.logBuffer.Truncate(0)
 	return nil
 }
